+++ CHANGELOG.md
@@ -27,6 +27,7 @@ delimited CGI params by default.
 - For pydantic schema based ``iemapp``, keys like `wfo[]` go to `wfo`.
 - Introduce some pydantic based validators for web requests, experimental...
 - Store `product_signature` in sbw table (requires iem-database schema update).
+- Support SPC corrections for MCDs so to prevent database dups.
 
 ### Bug Fixes